Simplistic user-system with full MML-Support for characters
A simplistic User-System allowing authorization, usernames and different MML-Characters with (permissioned) inventory for global and unique items.
This is easily adapted to centralized databases and NFT-based blockchain-applications.

Refer the User-System section in README.md for examples and details.

Technically
- Adds UserId which is hardcoded into the web-client
-
- A simplistic protocol with UserData message (for client-initated
  changes) and UserUpdate-message (for server-authorized distribution of
user-details such as username, character, ..)
- A simplistic inventory system supporting unique items, which can at
  most be used once within a 3d-web-experience.

Code structured in a way s.t. it can easily be adapted to actual
applications and authorization structures